+++ crates/rust-client/src/transaction/mod.rs
@@ -108,7 +108,6 @@ mod store_update;
 pub use store_update::TransactionStoreUpdate;
 
 mod request;
-use request::account_proof_into_inputs;
 pub use request::{
     ForeignAccount,
     NoteArgs,
@@ -118,6 +117,7 @@ pub use request::{
     TransactionRequestBuilder,
     TransactionRequestError,
     TransactionScriptTemplate,
+    account_proof_into_inputs,
 };
 
 mod result;
@@ -290,7 +290,7 @@ where
 
         let ignore_invalid_notes = transaction_request.ignore_invalid_input_notes();
 
-        let data_store = ClientDataStore::new(self.store.clone());
+        let data_store = ClientDataStore::new(self.store.clone(), self.rpc_api.clone());
         data_store.register_foreign_account_inputs(foreign_account_inputs.iter().cloned());
         for fpi_account in &foreign_account_inputs {
             data_store.mast_store().load_account_code(fpi_account.code());
@@ -471,7 +471,7 @@ where
 
         let account: Account = account_record.try_into()?;
 
-        let data_store = ClientDataStore::new(self.store.clone());
+        let data_store = ClientDataStore::new(self.store.clone(), self.rpc_api.clone());
 
         data_store.register_foreign_account_inputs(foreign_account_inputs.iter().cloned());
 
@@ -623,7 +623,7 @@ where
         tx_args: TransactionArgs,
     ) -> Result<InputNotes<InputNote>, ClientError> {
         loop {
-            let data_store = ClientDataStore::new(self.store.clone());
+            let data_store = ClientDataStore::new(self.store.clone(), self.rpc_api.clone());
 
             data_store.mast_store().load_account_code(account.code());
             let execution = NoteConsumptionChecker::new(&self.build_executor(&data_store)?)
@@ -704,6 +704,7 @@ where
                     known_account_code,
                 )
                 .await?;
+
             let foreign_account_inputs = match foreign_account {
                 ForeignAccount::Public(account_id, ..) => {
                     let foreign_account_inputs: AccountInputs =
@@ -721,7 +722,6 @@ where
                 },
                 ForeignAccount::Private(partial_account) => {
                     let (witness, _) = account_proof.into_parts();
-
                     AccountInputs::new(partial_account, witness)
                 },
             };